requirements.txt
# Existing dependencies from hybrid.py
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-groq
llama-index-retrievers-bm25

# Google API dependencies (for YouTube only)
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# OpenAI and Tavily
openai
tavily-python

# Other dependencies
python-dotenv
requests
torch
transformers
sentence-transformers
nltk
rank-bm25
mistralai

# File processing
python-magic
PyPDF2
PyMuPDF
docx2txt

# Logging and monitoring
structlog

# CLI dependencies
typer[all]
rich
click

# Groq
groq
